Added used EEPROM space.
Reordered Defines.h.
Use fixed sizes (uint16_t) instead of "unsigned int" in order to force
equal number of bytes on Desktop version and Arduboy version. Notably for saving
to and reading from EEPROM.
When saving to EEPROM, insert number of bytes used after the signature, to allow
appending data for another game after the first. Code converts from the old save format
automatically.
Makefile: Make build dependant on headers, too.

+++ 2048-Arduboy/2048-Arduboy.ino
@@ -109,7 +109,7 @@ unsigned long Millis() {
 
 #define EEP(x) EEPROM[EEPROM_STORAGE_SPACE_START + x]
 
-uint8_t Platform::ToEEPROM(uint8_t *bytes, int offset, int sz) {
+uint8_t Platform::ToEEPROM(uint8_t *bytes, int offset, uint16_t sz) {
   int i;
   if (offset < 0) {
     return WrongOffset;
@@ -125,7 +125,7 @@ uint8_t Platform::ToEEPROM(uint8_t *bytes, int offset, int sz) {
   return Saved;
 }
 
-uint8_t Platform::FromEEPROM(uint8_t *bytes, int offset, int sz) {
+uint8_t Platform::FromEEPROM(uint8_t *bytes, int offset, uint16_t sz) {
   // Get Game from EEPROM
   int i = EEPROM_STORAGE_SPACE_START + offset;
   if (i < 0) {
