+++ 2048-Arduboy/Platform.h
@@ -19,8 +19,8 @@ public:
   static void FillScreen(uint8_t colour);
   static void Clear();
 
-  static uint8_t ToEEPROM(uint8_t *bytes, int offset, int length);
-  static uint8_t FromEEPROM(uint8_t *bytes, int offset, int length);
+  static uint8_t ToEEPROM(uint8_t *bytes, int offset, uint16_t length);
+  static uint8_t FromEEPROM(uint8_t *bytes, int offset, uint16_t length);
 	/* static void PlaySound(const uint16_t* audioPattern); */
 	/* static bool IsAudioEnabled(); */
 	/* static void SetAudioEnabled(bool isEnabled); */
